+++ src/qrcode.ts
@@ -1,51 +1,14 @@
-import QRCode from 'qrcode';
+import qrcode from 'qrcode';
+import { createSetupUri } from './homekit';
 
-const generateHomeKitSetupUri = (
-  category: number = 0,
-  password: string = '',
-  setupId: string = '',
-  version = 0,
-  reserved: number = 0,
-  flags = 2
-): string => {
-  let payload: bigint | number = 0;
-  payload = payload | (version & 0x7);
-
-  payload = payload << 4;
-  payload = payload | (reserved & 0xf); // reserved bits
-
-  payload = payload << 8;
-  payload = payload | (category & 0xff);
-
-  payload = payload << 4;
-  payload = payload | (flags & 0xf);
-
-  payload = BigInt(payload) << BigInt(27);
-  payload =
-    BigInt(payload) | BigInt(Number(password.replace(/-/g, '')) & 0x7fffffff);
-
-  const payloadBase36 = payload.toString(36).toUpperCase().padStart(9, '0');
-
-  return `X-HM://${payloadBase36}${setupId}`;
+const QR_CODE_STRING_OPTIONS: qrcode.QRCodeToStringOptions = {
+  errorCorrectionLevel: 'quartile',
+  margin: 0,
+  type: 'svg',
+  version: 2,
 };
 
-export const generateQrCodeAsSvg = async (
-  category: number = 0,
-  password: string = '',
-  setupId = ''
-): Promise<string> => {
-  const svg = await QRCode.toString(
-    generateHomeKitSetupUri(category, password, setupId),
-    {
-      errorCorrectionLevel: 'Q',
-      margin: 0,
-      type: 'svg',
-      version: 2,
-    }
-  ).then((value) => {
-    // Add an id to the svg output
-    return value.replace(/<svg/, '<svg id="qrCode"');
-  });
-
-  return svg;
+export const generateQrCodeAsSvg = async (category: number, password: string, setupId: string): Promise<string> => {
+  const setupUri = createSetupUri(category, password, setupId);
+  return qrcode.toString(setupUri, QR_CODE_STRING_OPTIONS);
 };